requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
configobj==5.0.6
fastapi==0.82.0
feedgen==0.9.0
Genshi==0.7.7
h11==0.13.0
idna==3.3
lxml==4.9.1
pydantic==1.10.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
